# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lumi",
    "description": "AI image and video creative workspace by Lumisene",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/editor/media": {
      "get": {
        "tags": [
          "editor"
        ],
        "summary": "Media",
        "operationId": "media_api_v1_projects__project_id__editor_media_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/editor/documents": {
      "get": {
        "tags": [
          "editor"
        ],
        "summary": "Documents",
        "operationId": "documents_api_v1_projects__project_id__editor_documents_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Create",
        "operationId": "create_api_v1_projects__project_id__editor_documents_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocument"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/editor/documents/{document_id}": {
      "get": {
        "tags": [
          "editor"
        ],
        "summary": "Get",
        "operationId": "get_api_v1_projects__project_id__editor_documents__document_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "editor"
        ],
        "summary": "Save",
        "operationId": "save_api_v1_projects__project_id__editor_documents__document_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDocument"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/editor/documents/{document_id}/renders": {
      "post": {
        "tags": [
          "editor"
        ],
        "summary": "Export",
        "operationId": "export_api_v1_projects__project_id__editor_documents__document_id__renders_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/editor/renders/{task_id}": {
      "get": {
        "tags": [
          "editor"
        ],
        "summary": "Task",
        "operationId": "task_api_v1_projects__project_id__editor_renders__task_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/editor/documents/{document_id}/latest-render": {
      "get": {
        "tags": [
          "editor"
        ],
        "summary": "Latest Render",
        "operationId": "latest_render_api_v1_projects__project_id__editor_documents__document_id__latest_render_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RenderResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Latest Render Api V1 Projects  Project Id  Editor Documents  Document Id  Latest Render Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Clip": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "source_start_frame": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Source Start Frame"
          },
          "source_end_frame": {
            "type": "integer",
            "maximum": 18000.0,
            "exclusiveMinimum": 0.0,
            "title": "Source End Frame"
          },
          "original_duration": {
            "type": "integer",
            "maximum": 18000.0,
            "exclusiveMinimum": 0.0,
            "title": "Original Duration"
          },
          "timeline_start_frame": {
            "type": "integer",
            "maximum": 17999.0,
            "minimum": 0.0,
            "title": "Timeline Start Frame"
          },
          "duration": {
            "type": "integer",
            "maximum": 18000.0,
            "exclusiveMinimum": 0.0,
            "title": "Duration"
          },
          "volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "default": 1
          },
          "muted": {
            "type": "boolean",
            "title": "Muted",
            "default": false
          },
          "speed": {
            "type": "integer",
            "const": 1,
            "title": "Speed",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "asset_id",
          "source_start_frame",
          "source_end_frame",
          "original_duration",
          "timeline_start_frame",
          "duration"
        ],
        "title": "Clip"
      },
      "Composition": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "fps": {
            "type": "integer",
            "const": 30,
            "title": "Fps",
            "default": 30
          },
          "width": {
            "type": "integer",
            "multipleOf": 2.0,
            "maximum": 1920.0,
            "minimum": 240.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "multipleOf": 2.0,
            "maximum": 1920.0,
            "minimum": 240.0,
            "title": "Height"
          },
          "duration_in_frames": {
            "type": "integer",
            "maximum": 18000.0,
            "minimum": 1.0,
            "title": "Duration In Frames"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/VideoTrack"
            },
            "type": "array",
            "maxItems": 1,
            "minItems": 1,
            "title": "Tracks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "width",
          "height",
          "duration_in_frames",
          "tracks"
        ],
        "title": "Composition"
      },
      "CreateDocument": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "default": "未命名剪辑"
          },
          "composition": {
            "$ref": "#/components/schemas/Composition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "composition"
        ],
        "title": "CreateDocument"
      },
...
      "DocumentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DocumentList"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "composition": {
            "$ref": "#/components/schemas/Composition"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "composition",
          "version",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MediaResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MediaList"
      },
      "MediaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "width",
          "height"
        ],
        "title": "MediaResponse"
      },
...
      "RenderRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "RenderRequest"
      },
      "RenderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status"
          },
          "output_asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Asset Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "document_id",
          "status",
          "output_asset_id",
          "error"
        ],
        "title": "RenderResponse"
      },
      "SaveDocument": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "composition": {
            "$ref": "#/components/schemas/Composition"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "composition",
          "expected_version"
        ],
        "title": "SaveDocument"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoTrack": {
        "properties": {
          "id": {
            "type": "string",
            "const": "video",
            "title": "Id",
            "default": "video"
          },
          "type": {
            "type": "string",
            "const": "video",
            "title": "Type",
            "default": "video"
          },
          "clips": {
            "items": {
              "$ref": "#/components/schemas/Clip"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Clips"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "clips"
        ],
        "title": "VideoTrack"
      }
    }
  }
}